requirements.txt
pytest
python-dateutil
pytest-json-report
